+++ end_to_end_test/lib/values.g.dart
@@ -636,6 +636,8 @@ class _$CompoundValueExplicitNoNestingSerializer
   CompoundValueExplicitNoNesting deserialize(
       Serializers serializers, Iterable<Object> serialized,
       {FullType specifiedType = FullType.unspecified}) {
+    T $cast<T>(dynamic any) => any as T;
+
     final result = new CompoundValueExplicitNoNestingBuilder();
 
     final iterator = serialized.iterator;
@@ -645,8 +647,14 @@ class _$CompoundValueExplicitNoNestingSerializer
       final Object value = iterator.current;
       switch (key) {
         case 'simpleValue':
-          result.simpleValue.replace(serializers.deserialize(value,
-              specifiedType: const FullType(SimpleValue)) as SimpleValue);
+          var maybeBuilder = result.simpleValue;
+          var fieldValue = serializers.deserialize(value,
+              specifiedType: const FullType(SimpleValue)) as SimpleValue;
+          if (maybeBuilder == null) {
+            result.simpleValue = $cast(fieldValue.toBuilder());
+          } else {
+            maybeBuilder.replace(fieldValue);
+          }
           break;
         case 'validatedValue':
           result.validatedValue = serializers.deserialize(value,